refactor: updated expansion to fit refactored version

+++ src/main/java/nl/saxion/expansion/model/SLAPrinter.java
@@ -4,8 +4,8 @@ import nl.saxion.expansion.model.visitor.PrinterVisitor;
 
 public class SLAPrinter implements Printer {
     @Override
-    public int getId() {
-        return 0;
+    public Long getId() {
+        return 0L;
     }
 
     @Override
